requirements.txt
boto3>=1.26.0
botocore>=1.29.0
pyaudio>=0.2.11
python-dotenv>=0.19.0
amazon-transcribe>=0.6.0
asyncio
